+++ src/Markdown/Extension/Figure/Renderer/FigureRenderer.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Jidaikobo\Markdown\Extension\Figure\Renderer;
 
 use Jidaikobo\Markdown\Extension\Figure\Node\Figure;
+use Jidaikobo\Markdown\HtmlAttributes;
 use League\CommonMark\Node\Node;
 use League\CommonMark\Renderer\ChildNodeRendererInterface;
 use League\CommonMark\Renderer\NodeRendererInterface;
@@ -19,7 +20,7 @@ final class FigureRenderer implements NodeRendererInterface
 
         return new HtmlElement(
             'figure',
-            $node->data->get('attributes'),
+            HtmlAttributes::fromNode($node),
             $separator . $childRenderer->renderNodes($node->children()) . $separator
         );
     }